+++ fleetpings/static/fleetpings/libs/bootstrap5-autocomplete/1.1.39/autocomplete.js
@@ -22,6 +22,13 @@
  */
 
 /**
+ * @callback ValueCallback
+ * @param {String} value
+ * @param {Autocomplete} inst
+ * @returns {void}
+ */
+
+/**
  * @callback ServerCallback
  * @param {Response} response
  * @param {Autocomplete} inst
@@ -73,15 +80,16 @@
  * @property {String} datalist The id of the source datalist
  * @property {String} server Endpoint for data provider
  * @property {String} serverMethod HTTP request method for data provider, default is GET
- * @property {String|Object} serverParams Parameters to pass along to the server. You can specify a "related" key with the id of a related field.
+ * @property {String|Object} serverParams Parameters to pass along to the server.  You can specify a "related" key with (a) the id of a related field or (b) an array of related field ids.
  * @property {String} serverDataKey By default: data
- * @property {Object} fetchOptions Any other fetch options (https://developer.mozilla.org/en-US/docs/Web/API/fetch#syntax)
+ * @property {RequestInit} fetchOptions Any other fetch options (https://developer.mozilla.org/en-US/docs/Web/API/fetch#syntax)
  * @property {Boolean} liveServer Should the endpoint be called each time on input
  * @property {Boolean} noCache Prevent caching by appending a timestamp
  * @property {Number} debounceTime Debounce time for live server
  * @property {String} notFoundMessage Display a no suggestions found message. Leave empty to disable
  * @property {RenderCallback} onRenderItem Callback function that returns the label
  * @property {ItemCallback} onSelectItem Callback function to call on selection
+ * @property {ValueCallback} onClearItem Callback function to call on clear
  * @property {ServerCallback} onServerResponse Callback function to process server response. Must return a Promise
  * @property {ErrorCallback} onServerError Callback function to process server errors.
  * @property {ItemCallback} onChange Callback function to call on change-event. Returns currently selected item if any
@@ -133,6 +141,7 @@ const DEFAULTS = {
     return label;
   },
   onSelectItem: (item, inst) => {},
+  onClearItem: (item, inst) => {},
   onServerResponse: (response, inst) => {
     return response.json();
   },
@@ -158,6 +167,9 @@ const SHOW_CLASS = "show";
 const NEXT = "next";
 const PREV = "prev";
 
+/**
+ * @type WeakMap<HTMLElement, Autocomplete>
+ */
 const INSTANCE_MAP = new WeakMap();
 let counter = 0;
 let activeCounter = 0;
@@ -216,7 +228,7 @@ function fuzzyMatch(str, lookup) {
   let pos = 0;
   for (let i = 0; i < lookup.length; i++) {
     const c = lookup[i];
-    if (c == " ") continue;
+    if (c === " ") continue;
     pos = str.indexOf(c, pos) + 1;
     if (pos <= 0) {
       return false;
@@ -239,7 +251,7 @@ function insertAfter(el, newEl) {
  * @returns {string}
  */
 function decodeHtml(html) {
-  var txt = document.createElement("textarea");
+  const txt = document.createElement("textarea");
   txt.innerHTML = html;
   return txt.value;
 }
@@ -277,7 +289,7 @@ function nested(str, obj = "window") {
 class Autocomplete {
   /**
    * @param {HTMLInputElement} el
-   * @param {Config|Object} config
+   * @param {Partial<Config>} config
    */
   constructor(el, config = {}) {
     if (!(el instanceof HTMLElement)) {
@@ -329,7 +341,7 @@ class Autocomplete {
   /**
    * Attach to all elements matched by the selector
    * @param {string} selector
-   * @param {Config|Object} config
+   * @param {Partial<Config>} config
    */
   static init(selector = "input.autocomplete", config = {}) {
     /**
@@ -343,6 +355,7 @@ class Autocomplete {
 
   /**
    * @param {HTMLInputElement} el
+   * @returns {Autocomplete | null}
    */
   static getInstance(el) {
     return INSTANCE_MAP.has(el) ? INSTANCE_MAP.get(el) : null;
@@ -350,7 +363,8 @@ class Autocomplete {
 
   /**
    * @param {HTMLInputElement} el
-   * @param {Object} config
+   * @param {Partial<Config>} config
+   * @returns {Autocomplete}
    */
   static getOrCreateInstance(el, config = {}) {
     return this.getInstance(el) || new this(el, config);
@@ -407,7 +421,7 @@ class Autocomplete {
   };
 
   /**
-   * @param {Config|Object} config
+   * @param {Partial<Config>} config
    */
   _configure(config = {}) {
     this._config = Object.assign({}, DEFAULTS);
@@ -694,10 +708,14 @@ class Autocomplete {
   }
 
   clear() {
+	const v = this._searchInput.value;
+
     this._searchInput.value = "";
     if (this._hiddenInput) {
       this._hiddenInput.value = "";
     }
+
+    this._config.onClearItem(v, this);
   }
 
   // #endregion
@@ -941,11 +959,30 @@ class Autocomplete {
   }
 
   /**
+   * Get the active element, drilling into shadowRoot if necessary.
+   * @link https://www.abeautifulsite.net/posts/finding-the-active-element-in-a-shadow-root/
+   * @param {Document | ShadowRoot} root
+   * @returns {Element}
+   */
+  _getActiveElement(root = document) {
+    const activeEl = root.activeElement;
+
+    if (!activeEl) {
+      return null;
+    }
+
+    if (activeEl.shadowRoot) {
+      return this._getActiveElement(activeEl.shadowRoot);
+    }
+    return activeEl;
+  }
+
+  /**
    * Show drop menu with suggestions
    */
   _showSuggestions() {
     // It's not focused anymore
-    if (document.activeElement != this._searchInput) {
+    if (this._getActiveElement() !== this._searchInput) {
       return;
     }
     const lookup = normalize(this._searchInput.value);
@@ -959,8 +996,8 @@ class Autocomplete {
       // Check search length since we can trigger dropdown with arrow
       const showAllSuggestions = this._config.showAllSuggestions || lookup.length === 0;
       // Do we find a matching string or do we display immediately ?
-      let isMatched = lookup.length == 0 && this._config.suggestionsThreshold === 0;
-      if (!showAllSuggestions && lookup.length > 0) {
+      let isMatched = lookup.length === 0 && this._config.suggestionsThreshold === 0;
+      if (lookup.length > 0) {
         // match on any field
         this._config.searchFields.forEach((sf) => {
           const text = normalize(entry[sf]);
@@ -976,6 +1013,7 @@ class Autocomplete {
           }
         });
       }
+
       const selectFirst = isMatched || lookup.length === 0;
       if (showAllSuggestions || isMatched) {
         count++;
@@ -1083,9 +1121,8 @@ class Autocomplete {
    * Position the dropdown menu
    */
   _positionMenu() {
-    const styles = window.getComputedStyle(this._searchInput);
     const bounds = this._searchInput.getBoundingClientRect();
-    const isRTL = styles.direction === "rtl";
+    const isRTL = this._searchInput.dir === "rtl" || (this._searchInput.dir === "" && document.dir === "rtl");
     const fullWidth = this._config.fullWidth;
     const fixed = this._config.fixed;
 
@@ -1228,6 +1265,9 @@ class Autocomplete {
     }
     this._abortController = new AbortController();
 
+    // call early so the config can be changed dynamically
+    this._config.onBeforeFetch(this);
+
     // Read data params dynamically as well
     let extraParams = this._searchInput.dataset.serverParams || {};
     if (typeof extraParams == "string") {
@@ -1240,20 +1280,24 @@ class Autocomplete {
     if (this._config.noCache) {
       params.t = Date.now();
     }
-    // We have a related field
+
+     // We have a related field or an array of related fields
     if (params.related) {
-      /**
-       * @type {HTMLInputElement}
-       */
-      //@ts-ignore
-      const input = document.getElementById(params.related);
-      if (input) {
-        params.related = input.value;
-        const inputName = input.getAttribute("name");
-        if (inputName) {
-          params[inputName] = input.value;
+      // Check if params.related is an array
+      const relatedItems = Array.isArray(params.related) ? params.related : [params.related];
+
+      relatedItems.forEach((related) => {
+        const input = document.getElementById(related);
+        if (input) {
+          const inputValue = input.value;
+          const inputName = input.getAttribute("name");
+
+          // Update params with the input value
+          if (inputName) {
+            params[inputName] = inputValue;
+          }
         }
-      }
+      });
     }
 
     const urlParams = new URLSearchParams(params);
@@ -1266,11 +1310,12 @@ class Autocomplete {
     if (fetchOptions.method === "POST") {
       fetchOptions.body = urlParams;
     } else {
-      url += "?" + urlParams.toString();
+      if (url.indexOf('?') == -1) url += "?";
+      else url += "&";
+      url += urlParams.toString();
     }
 
     this._searchInput.classList.add(LOADING_CLASS);
-    this._config.onBeforeFetch(this);
 
     fetch(url, fetchOptions)
       .then((r) => this._config.onServerResponse(r, this))