+++ CHANGELOG.md
@@ -39,6 +39,10 @@ Section Order:
 
 - Error handling for Fetch API calls
 
+### Changed
+
+- JS Autocomplete library updated
+
 ## [3.5.4] - 2025-04-09
 
 ### Added